# fix data race related to VersionExists