@ValarDragon 2022-03-22
Oh wow great catch, version Exists was writing during the read lock! Ok sweet, LGTM, could fix the problem! Lets try to get into a release with the snapshot manager ASAP?