@p0mvn 2022-03-22
Yes, updating the SDK side right now. The rest of the comments are addressed